Got the dry-run commands working

+++ subtitler/subtitler.py
@@ -5,9 +5,10 @@
 import argparse
 import base64
 import json
+import math
 import os
 import subprocess
-from typing import List
+from typing import (Any, Dict, List, Tuple)
 import urllib.parse
 
 
@@ -28,17 +29,42 @@ def youtubeVideoID(url: str) -> str:
   return data['v'][0]
 
 
-def mysystem(command: List[str]) -> int:
-  if True:
-    print(command)
+def mysystem_wrapper(dry_run: bool, command: List[str]) -> int:
+  if dry_run:
+    print(" ".join(command))
     return 0
   res = subprocess.run(command)
   return res.returncode
 
 
-def run(url: str, filename: str, lang: str) -> None:
+# https://wiki.videolan.org/SubViewer/
+def timeize(seconds: float) -> str:
+  hours = int(math.floor(seconds / 3600))
+  seconds -= hours * 3600
+  minutes = int(math.floor(seconds / 60))
+  seconds -= minutes * 60
+  return "{hours:02d}:{minutes:02d}:{seconds:02.03f}".format(
+    hours=hours,
+    minutes=minutes,
+    seconds=seconds
+  )
+
+
+def textize(index_result_tuple: Tuple[int, Dict[str, Any]]) -> str:
+  index, result = index_result_tuple
+  end = timeize(float(result['end_time']))
+  start = timeize(float(result['start_time']))
+  text = result['alternatives'][0]['content']
+  return """{index}
+{start} --> {end}
+{text}""".format(index=index + 1, start=start, end=end, text=text)
+
+
+def gen_subtitles(url: str, filename: str, lang: str, dry_run: bool) -> None:
   bucket = 'subtitler1'
   region = 'us-east-2'
+
+  mysystem = lambda command: mysystem_wrapper(dry_run, command)
 
   job_id = urandom5()
   video_id = youtubeVideoID(url)
@@ -46,14 +72,20 @@ def run(url: str, filename: str, lang: str) -> None:
     job_id=job_id,
     video_id=video_id
   )
+  print('Running job_name: {job_name}'.format(job_name=job_name))
 
   _resp = mysystem(['youtube-dl', url, '--output', 'downloads/{video_id}.%(ext)s'.format(
     video_id=video_id
   )])
 
   files = [f for f in os.listdir('downloads') if f.startswith(video_id)]
+  if dry_run and len(files) == 0:
+    files = ['{video_id}.dummy.mkv'.format(video_id=video_id)]
   assert len(files) > 0, 'Expected at least one video file like %s.*' % video_id
   video_file = files[0]
+  file_parts = video_file.split('.')
+  assert len(file_parts) >= 2, 'Expected at least two parts in the video file name'
+  file_ext = file_parts[-1]
 
   # supported file types: mp3 | mp4 | wav | flac
   # from https://docs.aws.amazon.com/transcribe/latest/dg/API_TranscriptionJob.html#transcribe-Type-TranscriptionJob-MediaFormat
@@ -78,17 +110,20 @@ def run(url: str, filename: str, lang: str) -> None:
     'LanguageCode': lang,
     'MediaFormat': 'wav',
     'Media': {
-      'MediaFileUri': 'https://{bucket}.s3.{region}.amazonaws.com/{filename}'.format(
+      'MediaFileUri': 'https://{bucket}.s3.{region}.amazonaws.com/{video_id}.wav'.format(
         bucket=bucket,
-        filename=filename,
+        video_id=video_id,
         region=region,
       ),
     },
   }
   json_job_str = json.dumps(job_obj)
-  with open('job-start-command.json', 'wb') as f:
-    f.write(json_job_str.encode('utf-8'))
-    f.write(b'\n')
+  if not dry_run:
+    with open('job-start-command.json', 'wb') as f:
+      f.write(json_job_str.encode('utf-8'))
+      f.write(b'\n')
+  else:
+    print('job-start-command.json: {json_job_str}'.format(json_job_str=json_job_str))
 
   _resp = mysystem([
     'aws',
@@ -113,25 +148,50 @@ def run(url: str, filename: str, lang: str) -> None:
   ])
 
   # TODO split into two commands
-  _resp = os.system('curl -o output.json "$(aws transcribe get-transcription-job --region {region} --transcription-job-name {job_name} | jq -r .TranscriptionJob.Transcript.TranscriptFileUri)"'.format(job_name=job_name, region=region))
+  command = 'curl -o outputs/{video_id}.json "$(aws transcribe get-transcription-job --region {region} --transcription-job-name {job_name} | jq -r .TranscriptionJob.Transcript.TranscriptFileUri)"'.format(
+    job_name=job_name,
+    region=region,
+    video_id=video_id
+  )
+  if dry_run:
+    print(command)
+  else:
+    _resp = os.system(command)
 
   # jq -Cc '.results.items[]' output.json | head
-  # jq -cr '.results.items[] | [.start_time, .end_time, .alternatives[0].content] | @tsv' output.json | head
-  # TODO format the SUB file based on the JSON output
-  # https://wiki.videolan.org/SubViewer/
-  sub_format = """
-  [INFORMATION]
-  [TITLE] {video_title}
-  [END INFORMATION]
+  # jq -cr '.results.items[] | select(.start_time != null) | [.start_time, .end_time, .alternatives[0].content] | @tsv' output.json | head
 
-  00:00:00.00,00:00:10.00
-  some text
-  """ # hours:minutes:seconds.centiseconds
-  # timestart,timeend
-  # TODO ffmpeg -i {filename} -i formatted.srt -c copy -c:s mov_text final_{filename}
-  print('ffmpeg -i downloads/{video_file} -i formatted.srt -c copy -c:s mov_text final/{filename}')
+  with open('outputs/{video_id}.json'.format(video_id=video_id), 'rb') as f:
+    result_obj = json.loads(f.read().decode('utf-8'))
+
+  try:
+    results = [item for item in result_obj['results']['items']
+               if 'start_time' in item]
+  except KeyError as e:
+    print('%s: %s' % (type(e).__name__, str(e)))
+    return
+
+  subtitle_file = 'subtitles/{video_id}.srt'.format(video_id=video_id)
+  text = "\n\n".join(list(map(textize, enumerate(results))))
+  with open(subtitle_file, 'wb') as f:
+    f.write(text.encode('utf-8'))
+
+  _resp = mysystem([
+    'ffmpeg',
+    '-i',
+    'downloads/{video_file}'.format(video_file=video_file),
+    '-i',
+    subtitle_file,
+    '-s',
+    '720x480',
+    # '-c',
+    # 'copy',
+    '-c:s',
+    'mov_text',
+    'final/{filename}.mp4'.format(filename=filename),
+  ])
   return
-  # end def run
+  # end def gen_subtitles
 
 
 # `aws configure list` should show an access key and secret key. Else try
@@ -149,14 +209,17 @@ def main() -> None:
   parser.add_argument('video_url')
   parser.add_argument('temp_video_file')
   parser.add_argument('language')
+  parser.add_argument('--dry-run', action='store_true', help='Only print commands ' +
+                      'that would have run')
   args = parser.parse_args()
   language = args.language.lower()
   assert language in ['english', 'hindi']
   lang = 'en-US'
   if language == 'hindi':
     lang = 'hi-IN'
-  run(args.video_url, args.temp_video_file, lang)
+  gen_subtitles(args.video_url, args.temp_video_file, lang, args.dry_run)
   return
+  # end def main
 
 
 if __name__ == '__main__':
